docker/supervisor/requirements.txt
fastapi
uvicorn
requests
openai
pydantic
pillow
beeai-framework
